x-way/requirements.txt
django==3.2.3
whitenoise==5.2.0
python-decouple==3.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
celery==5.0.5
gunicorn==20.1.0
psycopg2-binary==2.8.6
selenium==3.141.0
pillow
